# docs/openapi.yaml
openapi: 3.1.0
info:
  title: baifa-api
  description: >
    Especificación OpenAPI de la API RESTful **baifa-api**.
    Desarrollada con Laravel 12 y Docker Compose (Nginx, PHP 8.4-FPM y MySQL 8.0).
    Incluye autenticación mediante Bearer Tokens con **Laravel Sanctum** y control de roles:
    Cliente (`client`), Empleado (`employee`), Jefe/Gerente (`manager`) y Administrador (`admin`).
  version: 1.0.3
  contact:
    name: Juan Chirinos
    email: juaneliezer13@gmail.com
servers:
  - url: http://localhost:8000/api
    description: Entorno de Desarrollo Local (Docker)
...
paths:
...
  /v1/tracking/{serial_number}:
    get:
      tags:
        - Tracking / Checkpoints
      summary: Consulta de trazabilidad de un generador por serial
      description: >
        Obtiene la ficha de un generador y su historial cronológico de puntos de control
        a partir de su serial de fábrica. Permite a clientes consultar exclusivamente sus
        propios generadores (RN-PRT-02, RN-TRK-01).
      operationId: trackBySerial
      security:
        - bearerAuth: []
      parameters:
        - name: serial_number
          in: path
          required: true
          description: Serial de fábrica único del equipo (ej. GEN-2026-0041).
          schema:
            type: string
            example: GEN-2026-0041
      responses:
        '200':
          description: Información del generador y bitácora de trazabilidad.
          content:
            application/json:
              schema:
                type: object
                properties:
                  generator:
                    $ref: '#/components/schemas/Generator'
        '403':
          description: Usuario con rol cliente no autorizado a consultar este equipo.
        '404':
          description: Generador no encontrado con el serial suministrado.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
      description: Introduce tu token Bearer emitido en /login o /register (ej. `Bearer 1|...`)
  schemas:
...
    GeneratorStatus:
      type: string
      enum:
        - warehouse
        - in_transit
        - checkpoint
        - delivered
        - installed
      description: |
        Estado del ciclo logístico del generador:
        * `warehouse`: En Almacén.
        * `in_transit`: En Tránsito hacia destino.
        * `checkpoint`: En Punto de Control / Alcabala.
        * `delivered`: Entregado en locación del cliente.
        * `installed`: Instalado y Operativo.
      example: warehouse
    Generator:
      type: object
      required:
        - id
        - serial_number
        - model
        - status
        - status_label
        - status_color
      properties:
        id:
          type: integer
          example: 1
        serial_number:
          type: string
          example: GEN-2026-0041
        name:
          type: string
          nullable: true
          example: Planta Auxiliar
        model:
          type: string
          example: Cummins 150 kVA Silent Diésel
        capacity_kva:
          type: number
          nullable: true
          example: 150.0
        status:
          $ref: '#/components/schemas/GeneratorStatus'
        status_label:
          type: string
          example: En Almacén
        status_color:
          type: string
          example: violet
        client_id:
          type: integer
          nullable: true
          example: 2
        client:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            company_fiscal_name:
              type: string
            company_short_name:
              type: string
            rif:
              type: string
        estimated_arrival_date:
          type: string
          format: date
          nullable: true
          example: '2026-10-05'
        photo_path:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        checkpoints:
          type: array
          items:
            $ref: '#/components/schemas/Checkpoint'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Checkpoint:
      type: object
      required:
        - id
        - generator_id
        - status
        - status_label
        - status_color
        - checkpoint_name
        - changed_by
      properties:
        id:
          type: integer
          example: 1
        generator_id:
          type: integer
          example: 1
        status:
          $ref: '#/components/schemas/GeneratorStatus'
        status_label:
          type: string
          example: En Tránsito
        status_color:
          type: string
          example: sky
        checkpoint_name:
          type: string
          example: Alcabala Guacara
        event_date:
          type: string
          example: '2026-08-30 07:30'
        event_date_iso:
          type: string
          format: date-time
          example: '2026-08-30T07:30:00Z'
        notes:
          type: string
          nullable: true
          example: 'Generador despachado desde warehouse Caracas. Transportista: Express Cargo VE.'
        user_id:
          type: integer
          nullable: true
          example: 2
        changed_by:
          type: string
          example: Adriana Morales
        changed_by_role:
          type: string
          example: Administrador
        created_at:
          type: string
          format: date-time
...